+++ src/dungeon.h
@@ -1,5 +1,4 @@
 #pragma once
-#include <stddef.h>
 
 typedef enum {
     ROOM_EMPTY,
@@ -15,12 +14,12 @@ typedef struct {
 } room_t;
 
 typedef struct {
-    size_t current;
-    size_t total_rooms;
+    unsigned current;
+    unsigned total_rooms;
     room_t *rooms;
 } dungeon_t;
 
-void dungeon_init(dungeon_t *d, const size_t total, const room_type_t *rooms);
+void dungeon_init(dungeon_t *d, const unsigned total, const room_type_t *rooms);
 void dungeon_destroy(dungeon_t *d);
 bool dungeon_open_door(dungeon_t *d);
 room_t* dungeon_current_room(dungeon_t *d);
